V5.9 R8.3
SP Change - uHerdSync - CreateHerdSyncData - Added ReportInDays to the Health events to be synced to the server.
MK Change - DairyData - CheckFiles - Set all vaccination events that have a report in days of a week or more as modified so they up to the server again with the report in days - SP request.